move e2e to a dir under service and rename, dont make a module just run as test
