+++ .github/workflows/checks.yaml
@@ -194,10 +194,8 @@ jobs:
       - run: go run ./examples encrypt "Hello Virtru"
       - run: go run ./examples decrypt sensitive.txt.tdf
       - run: go run ./examples decrypt sensitive.txt.tdf | grep "Hello Virtru"
-      - name: create data for e2e tests
-        run: go run ./e2e create-data
-      - name: run e2e tests
-        run: go run ./e2e roundtrip-tests
+      - name: create roundtrip test data and run tests
+        run: go test ./service/rttests -v
 
   image:
     name: image build
