--- e2e/README.md
@@ -1,17 +1,0 @@
-# OpenTDF Platform E2E Tests
-
-This module contains roundtrip tests with different attribute types and combinations and a mechanism for creating test data.
-
-These tests can be launched using an e2e CLI. [See the autogenerated cobra docs](./docs/e2e.md).
-
-Example: Build and run the tests:
-```shell
-go build
-```
-
-```shell
-# create the test data
-./e2e create-data
-# run the tests
-./e2e roundtrip-tests
-```
