--- e2e/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/opentdf/platform/e2e/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
